+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 
 ### Removed
 - Remove `create`, `destroy`
+- Remove `TimerWheelDied` exception. `with` now simply re-throws the exception that the timer wheel thread throws.
 - Remove support for GHC < 8.6
 
 ## [0.2.0.1] - 2019-05-19